Rename util/cert-pin-format.ts -> util/pin-format.ts
Follow-up to PR #305 (which retitled the user-visible 'Cert
fingerprint' surface). PR #305 left the filename
'util/cert-pin-format.ts' alone to keep the diff focused on
cosmetic content changes; the rename naturally lands as a
small structural follow-up.

The function name (formatPinSha256) was already
shape-agnostic and stays. Test file renamed alongside.

Five importers updated:
- src/components/settings-dialog.ts
- src/components/accept-peer-dialog.ts
- src/components/pair-build-server-dialog.ts
- src/components/reauth-wizard-dialog.ts
- test/util/pin-format.test.ts

Verified rename tracking: git status shows R (rename),
not D + A, so blame history follows through.

1034 vitest tests pass; typecheck clean.
