+++ src/components/accept-peer-dialog.ts
@@ -7,7 +7,7 @@ import type { PeerSummary } from "../api/types.js";
 import { localizeContext } from "../context/index.js";
 import { pinHexStyles } from "../styles/pin-hex.js";
 import { espHomeStyles } from "../styles/shared.js";
-import { formatPinSha256 } from "../util/cert-pin-format.js";
+import { formatPinSha256 } from "../util/pin-format.js";
 import { registerMdiIcons } from "../util/register-icons.js";
 import "./confirm-dialog.js";
 import type { ESPHomeConfirmDialog } from "./confirm-dialog.js";
