+++ src/components/pair-build-server-dialog.ts
@@ -15,7 +15,7 @@ import { dialogActionButtonStyles } from "../styles/dialog-action-buttons.js";
 import { inputStyles } from "../styles/inputs.js";
 import { pinHexStyles } from "../styles/pin-hex.js";
 import { espHomeStyles } from "../styles/shared.js";
-import { formatPinSha256 } from "../util/cert-pin-format.js";
+import { formatPinSha256 } from "../util/pin-format.js";
 import {
   friendlyHostname,
   parsePortInput,
