+++ src/components/reauth-wizard-dialog.ts
@@ -8,7 +8,7 @@ import type { OffloaderPinMismatchAlert } from "../api/types.js";
 import { dialogActionButtonStyles } from "../styles/dialog-action-buttons.js";
 import { pinHexStyles } from "../styles/pin-hex.js";
 import { espHomeStyles } from "../styles/shared.js";
-import { formatPinSha256 } from "../util/cert-pin-format.js";
+import { formatPinSha256 } from "../util/pin-format.js";
 import { trimTrailingDot } from "../util/hostname.js";
 
 import "./base-dialog.js";
